Corregir code smells de severidad alta detectados por SonarCloud
- ResponseEntity<?> con comodin generico -> ResponseEntity<Object>
  (SalaRestController, InvitacionSalaController).
- Extraer constantes para strings repetidos: "error" (InvitacionSalaController,
  SalaRestController), "/topic/sala/" (JuegoWebSocketController),
  "codigo"/"estado"/"totalRondas" (SalaRestController, ahora armados en un
  unico metodo resumenBasico()).
- Sala.buscarJugador: fusionar ifs anidados.
- MotorJuego: quitar un cast a double redundante (la division ya promueve
  a double por la promocion numerica de Java).
- BancoPreguntasDemo: agregar constructor privado (clase de solo metodos
  estaticos).
- JuegoWebSocketController: quitar llaves innecesarias en un lambda de una
  sola instruccion.

Eran los hallazgos que hacian fallar el Quality Gate de "codigo nuevo"
(rating de mantenibilidad C). Verificado: las 63 pruebas siguen en verde.

+++ src/main/java/com/quizarena/juego/servicio/MotorJuego.java
@@ -33,7 +33,7 @@ public class MotorJuego {
             return PUNTOS_BASE;
         }
         // Factor de rapidez: 1.0 si responde al instante, baja hacia 0.5 al agotar el tiempo
-        double factorRapidez = 1.0 - (0.5 * Math.min(tiempoRespuesta, tiempoLimiteMs) / (double) tiempoLimiteMs);
+        double factorRapidez = 1.0 - (0.5 * Math.min(tiempoRespuesta, tiempoLimiteMs) / tiempoLimiteMs);
         return (int) Math.round(PUNTOS_BASE * factorRapidez);
     }
